update to command-line-args w/default option/value fix
command-line-args 4.0.1 used the default value for a default option
instead of one specified on the command line because of
<https://github.com/75lb/command-line-args/issues/47>, which we worked around
by setting the default value to the one specified on the command line.

command-line-args 4.0.3 fixes that problem, but in the process
it introduces a new one: if you specify the same value on the command line
as the default value, then command-line-args ignores the value
and sets the option to the next unknown value on the command line
<https://github.com/75lb/command-line-args/pull/48>.

Since our workaround makes the default value the same as the one specified
on the command line, it runs into that bug.  Fortunately, with 4.0.3,
we can stop setting the default value to the one specified on the command line,
resolving this problem (except in the rare event that the one specified
on the command line happens to be the same as the true default value, which is
the absolute path to the current working directory).

So this commit updates command-line-args to 4.0.3 and removes the workaround.
It also changes the test output to reflect that command-line-args no longer
includes the value of the default option in its list of "unknown args"
(which it should never have done but was previously doing, although I don't
know of an issue, PR, or commit that explicitly references or addresses it).