remove aliases and documentation of internal options
The --jsdebugger and --wait-for-jsdebugger options are intended
to be "internal" options for debugging the runtime itself, not public
options to be used by application developers.  So we should stop
advertising them to those developers.  This change does that.

In the process, it also removes the "aliases" (a.k.a. one-character shortcuts)
for those options and the --path option, since the jsdebugger options
aren't used frequently enough to justify burning aliases on them,
and the --path option isn't intended to be available at all (which is why
we don't document it).

Regarding --path: it's only an implementation detail of the way we implement
support for the path *argument*: we process arguments with command-line-args,
which supports arbitrary arguments by enabling you to specify a "default
option".  We specify --path as the default option so we can use that package
to read the argument.