return transformed result; log warnings to stdout; clarify stderr failure
Ensure the transform function in readProjectMetadata returns the result,
since the function that calls it expects to use its return value.

Log expected warnings to stdout rather than stderr, since the tests
automatically fail on output to stderr.  (Alternately, we could make the tests
expect those warnings.)

When tests fail because of output to stderr, make the cause of the failure
more obvious, logging the message 'test failure due to stderr output'.