remove run-url-bad-package-name.js as it doesn't test a bad package name
When you run a URL, qbrt uses the shell package to run it, and it has a good
package name.  So this test doesn't actually do what it's name suggests.
Nor is it necessary to test this, as run-path-bad-package.name.js is sufficient
to confirm that a bad package name will trigger a failure.