ensure result.pkg and result.path are defined; avoid finding qbrt's package.json
Ensure result.pkg and result.path are defined if we don't find a package.json
file, since the code assumes those properties are defined.

Avoid finding qbrt's package.json when traversing the directory tree
for the package.json file belonging to the test app that is missing that file.
This ensures that we actually test what happens when an app is missing
its package.json file, rather than testing what happeens when an app reuses
qbrt's package.json file (which, by the way, results in failure).