run tests as non-root user
npm sets TMPDIR to the cwd if running as the root user, per https://github.com/npm/npm/issues/4531, which is the, ahem, root cause of the test failure in run-path-missage-package.js.  There are two workarounds: invoke npm with --unsafe-perm or run it as a different user.  This commit attempts to do the latter.