wait for DevTools window to close before completing test
The DevTools windows tests get stuck on Mac in automation sometimes. I suspect it's because opening a DevTools window fails if another DevTools window is still open (even if it's in the process of unloading). So this change awaits the 'unload' event on each DevTools window before loading the next one.